+++ app/org/tronscan/importer/SolidityNodeImporter.scala
@@ -1,25 +1,21 @@
 package org.tronscan.importer
 
-import akka.actor.Actor
-import akka.stream._
-import akka.stream.scaladsl.{Flow, Keep, Sink, Source}
+import akka.actor.ActorSystem
+import akka.stream.scaladsl.{Keep, Source}
 import javax.inject.Inject
 import monix.execution.Scheduler.Implicits.global
-import org.tronscan.grpc.{SolidityBlockChain, WalletClient}
-import org.tronscan.importer.ImportManager.Sync
+import org.tronscan.grpc.WalletClient
 import org.tronscan.service.SynchronisationService
 import play.api.Logger
 
 import scala.async.Async.{async, await}
-import scala.concurrent.duration._
-import scala.util.{Failure, Success}
 
 
 class SolidityNodeImporter @Inject()(
   importersFactory: ImportersFactory,
   importStreamFactory: ImportStreamFactory,
   synchronisationService: SynchronisationService,
-  walletClient: WalletClient) extends Actor {
+  walletClient: WalletClient) {
 
 
   /**
@@ -32,14 +28,14 @@ class SolidityNodeImporter @Inject()(
     * 5. `synchronisationChecker`: verifies if the synchranisation should start
     * 6. `blockSource`: Builds the source from which the blocks will be read from the blockchain
     */
-  def buildStream() = {
-    implicit val system = context.system
+  def buildStream(implicit actorSystem: ActorSystem) = {
+
     async {
       val nodeState = await(synchronisationService.nodeState)
       Logger.info("BuildStream::nodeState -> " + nodeState)
       val importAction = await(importStreamFactory.buildImportActionFromImportStatus(nodeState))
       Logger.info("BuildStream::importAction -> " + importAction)
-      val importers = importersFactory.buildFullNodeImporters(importAction)
+      val importers = importersFactory.buildSolidityImporters(importAction)
       Logger.info("BuildStream::importers -> " + importers.debug)
       val synchronisationChecker = importStreamFactory.solidityNodePreSynchronisationChecker
       val blockSource = importStreamFactory.buildSolidityBlockSource(walletClient)
@@ -51,34 +47,5 @@ class SolidityNodeImporter @Inject()(
         .via(blockSource)
         .toMat(blockSink)(Keep.right)
     }
-  }
-
-  def startSync() = {
-
-    val decider: Supervision.Decider = { exc =>
-      Logger.error("FULL NODE ERROR", exc)
-      Supervision.Restart
-    }
-
-    implicit val materializer = ActorMaterializer(
-      ActorMaterializerSettings(context.system)
-        .withSupervisionStrategy(decider))(context)
-
-    Logger.info("START SOLIDITY NODE SYNC")
-
-    Source.tick(0.seconds, 2.8.seconds, "")
-      .mapAsync(1)(_ => buildStream().flatMap(_.run()))
-      .runWith(Sink.ignore)
-      .andThen {
-        case Success(_) =>
-          Logger.info("SOLIDITY SYNC SUCCESS")
-        case Failure(exc) =>
-          Logger.error("SOLIDITY SYNC FAILURE", exc)
-      }
-  }
-
-  def receive = {
-    case Sync() =>
-      startSync()
   }
 }
